fix:#01 unrecognized pragma
