+++ src/03_09_effect1/index.html
@@ -21,18 +21,12 @@
     </div>
     <div id="global-container">
       <div id="page-container">
-        <div id="partical-slide">
-          <div
-            class="div"
-            data-webgl="05partical-slide"
-            data-tex-1="/img/output0.jpg"
-            data-tex-2="/img/output1.jpg"
-          ></div>
-        </div>
-        <!-- <div class="div">
-          <img src="/img/output1.jpg" alt="" />
-        </div> -->
-
+         <div
+          id="div1"
+          data-webgl="06gray-pic"
+          data-tex-1="/img/output4.jpg"
+          data-tex-2="/img/output5.jpg"
+        ></div>
         <div
           id="div1"
           data-webgl="01sticky-plane"
@@ -58,12 +52,20 @@
           data-tex-2="/img/output5.jpg"
           data-tex-disp="/img/disps/3.png"
         ></div>
-        <!-- <div
-          id="div1"
-          data-webgl="05partical-slide"
-          data-tex-1="/img/output4.jpg"
-          data-tex-2="/img/output5.jpg"
-        ></div> -->
+
+        <div id="partical-slide">
+          <div
+            class="div"
+            data-webgl="05partical-slide"
+            data-tex-1="/img/output0.jpg"
+            data-tex-2="/img/output1.jpg"
+          ></div>
+        </div>
+        <!-- <div class="div">
+          <img src="/img/output1.jpg" alt="" />
+        </div> -->
+
+       
         <!-- <div
           id="div1"
           data-webgl="normal"
