+++ src/03_09_effect1/scripts/glsl/00_template/index.js
@@ -1,3 +1,5 @@
+import gsap from "gsap";
+
 import vertexShader from "./vertex.glsl";
 import fragmentShader from "./fragment.glsl";
