+++ src/03_09_effect1/scripts/glsl/00_template/vertex.glsl
