+++ src/03_09_effect1/scripts/glsl/06gray-pic/fragment.glsl
@@ -9,6 +9,8 @@ vec2 coverUv(vec2 uv, vec4 resolution) {
     return (uv - .5) * resolution.zw + .5;
 }
 
+#pragma glslify: require(../shader-helper/grayscale);
+
 void main() {
           // vec2 mouse = step(uMouse, vUv);
           // gl_FragColor = vec4(mouse, uHover, 1.);
@@ -18,5 +20,7 @@ void main() {
     vec4 t1 = texture2D(tex1, uv);
     vec4 t2 = texture2D(tex2, uv);
     vec4 color = mix(t1, t2, step(.5, uv.x));
-    gl_FragColor = t2;
+    vec4 gray = grayscale(color);
+    gl_FragColor = color;
+    gl_FragColor = gray;
 }
