+++ src/03_09_effect1/scripts/glsl/06gray-pic/index.js
@@ -1,0 +1,33 @@
+import gsap from "gsap";
+import vertexShader from "./vertex.glsl";
+import fragmentShader from "./fragment.glsl";
+
+import { CustomObject } from "../CustomObject";
+
+class ExtendObject extends CustomObject {
+  fixVertex() {
+    return vertexShader;
+  }
+
+  fixFragment() {
+    return fragmentShader;
+  }
+
+  debug(toFolder) {
+    toFolder
+      .add(this.uniforms.uProgress, "value", 0, 1, 0.1)
+      .name("progress")
+      .listen();
+
+    const datData = { next: !!this.uniforms.uProgress.value };
+    toFolder.add(datData, "next").onChange(() => {
+      gsap.to(this.uniforms.uProgress, {
+        value: +datData.next,
+        duration: 2,
+        ease: "power2.inOut",
+      });
+    });
+  }
+}
+
+export default ExtendObject;
