+++ src/03_09_effect1/scripts/glsl/06gray-pic/vertex.glsl
