+++ src/03_09_effect1/scripts/glsl/shader-helper/grayscale.glsl
@@ -1,0 +1,7 @@
+vec4 grayscale(vec4) {
+    vec3 gray = vec3(0.299, 0.587, 0.114);
+    float grayScale = dot(tex.rgb, gray);
+    return vec4(grayScale, grayScale, grayScale, tex.a);
+}
+
+#pragma glslify: export(vec4 grayscale(vec4))
