--- src/03_09_effect1/scripts/glsl/template/index.js
@@ -1,16 +1,0 @@
-import vertexShader from "./vertex.glsl";
-import fragmentShader from "./fragment.glsl";
-
-import { CustomObject } from "../CustomObject";
-
-class ExtendObject extends CustomObject {
-  fixVertex() {
-    return vertexShader;
-  }
-
-  fixFragment() {
-    return fragmentShader;
-  }
-}
-
-export default ExtendObject;
